ppp: Add BTRFS options to kernel config
I installed my PinePhone Pro with BTRFS as the root filesystem, which
involved the very changes I'm committing here. The reason for going with
BTRFS is because I use it on almost all machines and not only is it much
easier to send/receive snapshots, it also comes with a few other
benefits such as compression - which IMHO is especially useful on such a
device with constrained storage.

Signed-off-by: aszlig <aszlig@nix.build>